face_recognition/requirements.txt
CMake >= 3.12.0
dlib >= 19.7.0
face_recognition
opencv-python
Pillow
